% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueReport.R
\name{issuesData}
\alias{issuesData}
\title{Get Issues data from GitHub.}
\usage{
issuesData(issueNumber = 1, githubUsername = "xxxxxxxxxxxxxxxxxxxx",
  githubToken = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
  githubUrl = "https://api.github.com/",
  githubRepo = "mkirch/issueReporter")
}
\arguments{
\item{issueNumber}{Integer GitHub issue number.}

\item{githubUsername}{String of GitHub username to authenticate with.}

\item{githubToken}{String containing GitHub personal access token, as
generated from the settings area in your GitHub account.}

\item{githubUrl}{String containing the API URL for GitHub. If using a GitHub
Enterprise account, use http(s)://hostname/api/v3/ instead of default. Please
make sure to append a \code{`/`}.}

\item{githubRepo}{String of format \code{user/repo} specifying the GH repo.}
}
\value{
Returns a list of \code{list(title,body, c(users), c(nicknames), c(notes))}.
}
\description{
For a given GitHub issue, extracts all issue comments.
}
