% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issueReport.R
\name{issuesData2md}
\alias{issuesData2md}
\title{Convert GitHub Issues data to Markdown.}
\usage{
issuesData2md(issuesData = readRDS(system.file("extdata", "dataExample.rds",
  package = "issueReporter")), includeTitle = TRUE, includeBody = TRUE)
}
\arguments{
\item{issuesData}{The list output from \code{issuesData()} containing
GitHub Issues data.}

\item{includeTitle}{Bool where TRUE includes the title, and FALSE excludes.}

\item{includeBody}{Bool where TRUE includes the original issue, and FALSE
includes only the issue comments.}
}
\value{
Returns a character vector containing the markdown formatted text.
}
\description{
After being passed GitHub Issues data for a given thread, extracts and
formats the material in Markdown.
}
